requirements.txt
asgiref==3.7.2
Brotli==1.0.9
cffi==1.15.1
crispy-bootstrap5==0.7
cssselect2==0.7.0
Django==4.2.4
django-crispy-forms==2.0
django-filter==23.2
fonttools==4.42.1
html5lib==1.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.16.1
pycparser==2.21
pydyf==0.7.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
tinycss2==1.2.1
typing-extensions==4.7.1
tzdata==2023.3
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.5.0
zopfli==0.2.2
